backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
groq
python-multipart
